Update miniDMIReworked.ino
Got stuff mostly working how I want, I only need to get the little led and button board to do something like show the time, or reveal other information and such when a button is pressed. There's a problem where the list of calling points gets cut off after around 6 or 7 due to a lack of memory I think. Need to address this.